code/requirements.txt
requests
python-pptx
python-docx
pandas
openpyxl 